Rename style/__tests__/Text.test.js to TextStyle.test.js
Before this change, we had two test files called `Text.test.js`, in different folders in the project.

Depending on random factors, one could overwrite the other while compiling the test plugin, so we'd actually be running the tests for Text Styles when we think we should be running Text tests, and viceversa.

Renaming the Text Styles test file to `TextStyles.test.js` fixes this

I've been hitting my head against the desk, and I am _so_ going to have a cold beverage to reward me for this discovery.

This also fixes #748