Update tests to use same API as exposed by sketch module
Some tests imported modules relatively which are not available or must be imported differently when writing a script/plugin.

Given that all tests run as part of a plugin, they now use the exact same API as every other plugin. Relative imports are resolved as externals by the webpack configuration.