Use rawValue to get platform independent value
The API accepts both `NSTextAlignmentCenter` (which is the integer 2) or
`'center'` as an argument to `style.alignment`. We need to find the raw
numeric value before we can get the platform independent value.

The code assumes that whenever we cannot find the raw value whatever the
user passed in will be a correct value. This is what the previous
implementation did as well.