Remove incorrect doc comments
Both doc comments claim that reading and writing global settings is the
equivalent to using defaults read/write. For reading this is only true
when the stored value is not valid JSON.

Writing a global setting is never the same as `defaults write` because
we always store `JSON.stringify` the value before writing it to user
defaults.